% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_pcor_shrunk.R
\name{c_pcor_shrunk}
\alias{c_pcor_shrunk}
\title{Partial correlation shrunk}
\usage{
c_pcor_shrunk(lparams)
}
\arguments{
\item{lparams}{a list of parameters created using a JSON file.
This file should contain the following name/value pairs.

"filename": <string, required>

"variables": <array, strings representing column names>

"cutoff": <number, required threshold for the p-value of the partial correlation>

"verbose": <boolean, required to display detailed description on the terminal>}
}
\value{
Forest plot of partial correlations in Rplot.pdf
}
\description{
This function computes confidence intervals for the partial correlation with shrinkage.
}
