% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_pcor_shrunk.R
\name{c_zscore_shrunk}
\alias{c_zscore_shrunk}
\title{c_zscore_shrunk}
\usage{
c_zscore_shrunk(lparams)
}
\arguments{
\item{lparams}{a list of parameters created using a JSON file.
This file should contain the following name/value pairs.

"filename": <string, required data file 1>

"filename2": <string, required data file 2>

"variables": <array, strings representing column names>

"cutoff": <number, required threshold for the p-value of the partial correlation>

"verbose": <boolean, required to display detailed description on the terminal>}
}
\value{
a scatter plot
}
\description{
This function compares two networks using two data files
}
\details{
TODO: add proper title and description
}
