% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcor_Comparison.R
\docType{package}
\name{compare.GGM}
\alias{compare.GGM}
\title{compare.GGM}
\usage{
compare.GGM(x1 = NULL, x2 = NULL)
}
\arguments{
\item{x1}{data matrix}

\item{x2}{shrinkage value \code{lambda}}
}
\value{
z-score that compares two (partial) correlations accounting for the shrinkage \code{lambda}.
}
\description{
Test whether two (partial) correlations are equal.
}
\examples{
#---------------------------------
# 1- Simulate a 40 x 40 matrix of partial correlations
true.pcor <- GeneNet::ggm.simulate.pcor(40)

# 2- Simulate random data from true.pcor
dat.sim1 <- GeneNet::ggm.simulate.data(40, true.pcor)
dat.sim2 <- GeneNet::ggm.simulate.data(40, true.pcor)

# 3- Estimate the partial correlations with shrinkage
estimated.pcor1 <- GGM.shrunk(dat.sim1)
estimated.pcor2 <- GGM.shrunk(dat.sim2)

# 4- Compare
z <- compare.GGM(estimated.pcor1, estimated.pcor2)

# 5- Compare the estimate with the true value
sum( abs(z) >=1.96)
plot( abs(z), pch=20, ylim=c(0,4) , cex=0.5)
abline(h=1.96)
#---------------------------------

}
\references{
\url{to be submitted}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
