% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcor_CI.R
\docType{package}
\name{confint.GGM}
\alias{confint.GGM}
\title{confint.GGM}
\usage{
confint.GGM(x = NULL, alpha = 0.05)
}
\arguments{
\item{x}{vector of (partial) correlations}

\item{alpha}{significance level (two-sided test)}
}
\value{
Confidence intervals of the (partial) correlations with shrinkage \code{lambda}.
}
\description{
Confidence intervals for the (partial) correlation with shrinkage.
}
\examples{
#---------------------------------
# 1- Simulate a 40 x 40 matrix of partial correlations
true.pcor <- GeneNet::ggm.simulate.pcor(40)

# 2- Simulate random data from true.pcor
dat.sim <- GeneNet::ggm.simulate.data(40, true.pcor)

# 3- Estimate the partial correlations with shrinkage
estimated.pcor <- GGM.shrunk(dat.sim)

# 4- Compare the estimate with the true value
confint.GGM( x = estimated.pcor, alpha = 0.05 )

#---------------------------------

}
\references{
\url{to be submitted}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
