% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corr_shrunk.R
\docType{package}
\name{corr.shrunk}
\alias{corr.shrunk}
\title{corr.shrunk}
\usage{
corr.shrunk(x = NULL, lambda, verbose = TRUE)
}
\arguments{
\item{x}{data matrix}

\item{lambda}{shrinkage value}

\item{verbose}{print details}
}
\value{
The matrix of (Pearson's) correlations \code{R} with shrinkage.
}
\description{
Computes the (Pearson's) correlation matrix with shrinkage.
}
\references{
\url{https://doi.org/10.2202/1544-6115.1175}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
