% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcor_Shrunk.R
\docType{package}
\name{GGM.shrunk}
\alias{GGM.shrunk}
\title{GGM.shrunk}
\usage{
GGM.shrunk(x = NULL, lambda, verbose = TRUE)
}
\arguments{
\item{x}{data matrix}

\item{lambda}{shrinkage value (see \code{\link[optimal.shrinkage]{par}})}

\item{verbose}{print details}
}
\value{
The matrix of partial correlations \code{P} with shrinkage \code{lambda}.
}
\description{
Computes the partial correlations with shrinkage.
}
\examples{
#---------------------------------
}
\references{
\url{https://doi.org/10.2202/1544-6115.1175; to be submitted}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
