% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDfunc.R
\docType{package}
\name{id.func}
\alias{id.func}
\title{id.func}
\usage{
id.func(x)
}
\arguments{
\item{x}{correlation matrix}
}
\value{
names (identifiers) for the correlations
}
\description{
assign names.
}
\references{
\url{https://doi.org/10.2202/1544-6115.1175}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
