% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcor_DegreesFreedom.R
\docType{package}
\name{k.shrunk}
\alias{k.shrunk}
\title{k.shrunk}
\usage{
k.shrunk(p, n, lambda)
}
\arguments{
\item{p}{number of variables}

\item{n}{number of samples}

\item{lambda}{(optional) shrinkage value (see \code{\link[lambda]{par}})}
}
\value{
degrees of freedom with shrinkage \code{k}
}
\description{
Degrees of freedom k via Maximum Likelihood Estimation
}
\examples{
#---------------------------------
# 1- Simulate a 40 x 40 matrix of partial correlations
true.pcor <- GeneNet::ggm.simulate.pcor(40)

# 2- Simulate random data from true.pcor
dat.sim <- GeneNet::ggm.simulate.data(40, true.pcor)

# 3- Estimate the degrees of freedom
k.shrunk(p = ncol(dat.sim) , n = nrow(dat.sim) , lambda = 0.1)
#---------------------------------

}
\references{
\url{https://doi.org/10.2202/1544-6115.1175}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
