% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShrinkageOptimalMLE.R
\docType{package}
\name{optimal.shrinkage}
\alias{optimal.shrinkage}
\title{optimal.shrinkage}
\usage{
optimal.shrinkage(x = NULL)
}
\arguments{
\item{x}{data matrix}
}
\value{
optimal shrinkage value \code{lambda}.
}
\description{
Computes the optimal shrinkage value
}
\examples{
#---------------------------------
# 1- Simulate a 40 x 40 matrix of partial correlations
true.pcor <- GeneNet::ggm.simulate.pcor(40)

# 2- Simulate random data from true.pcor
dat.sim <- GeneNet::ggm.simulate.data(40, true.pcor)

# 3- Estimate the optimal shrinkage value
lambda <- optimal.shrinkage(x = dat.sim)
lambda
#---------------------------------

}
\references{
\url{https://doi.org/10.2202/1544-6115.1175}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
