% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pval_MonteCarlo.R
\docType{package}
\name{pval.montecarlo}
\alias{pval.montecarlo}
\title{pval.montecarlo}
\usage{
pval.montecarlo(r = NULL, number = 10)
}
\arguments{
\item{r}{matrix of partial correlations}

\item{number}{number of simulations}
}
\value{
vector of p-values for the partial correlations \code{P} with shrinkage \code{lambda}.
}
\description{
P-values for the partial correlations with shrinkage using MonteCarlo.
It test whether the partial correlation is different from zero.
}
\examples{
#---------------------------------
# 1- Simulate a 40 x 40 matrix of partial correlations
true.pcor <- GeneNet::ggm.simulate.pcor(40)

# 2- Simulate random data from true.pcor
dat.sim <- GeneNet::ggm.simulate.data(40, true.pcor)

# 3- Estimate the partial correlations with shrinkage
estimated.pcor <- GGM.shrunk(dat.sim)
pval <- pval.montecarlo(r = estimated.pcor)

#---------------------------------

}
\references{
\url{https://doi.org/10.1093/bioinformatics/btz357}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
