% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pval_Shrunk.R
\docType{package}
\name{pval.shrunk}
\alias{pval.shrunk}
\title{pval.shrunk}
\usage{
pval.shrunk(x = NULL, lambda = NULL)
}
\arguments{
\item{x}{matrix of partial correlations}

\item{lambda}{(optional) shrinkage value (see \code{\link[optimal.shrinkage]{par}})}
}
\value{
vector of p-values for the partial correlations \code{P} with shrinkage \code{lambda}.
}
\description{
P-values for the partial correlations with shrinkage.
Test whether the partial correlation is different from zero.
}
\examples{
#---------------------------------
# 1- Simulate a 40 x 40 matrix of partial correlations
true.pcor <- GeneNet::ggm.simulate.pcor(40)

# 2- Simulate random data from true.pcor
dat.sim <- GeneNet::ggm.simulate.data(40, true.pcor)

# 3- Estimate the partial correlations with shrinkage
estimated.pcor <- GGM.shrunk(dat.sim)

# 4.- P-values
pval <- pval.shrunk( x = estimated.pcor)

#---------------------------------

}
\references{
\url{https://doi.org/10.1093/bioinformatics/btz357}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
