% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2vecTools.R
\docType{package}
\name{sm2vec.tools}
\alias{sm2vec.tools}
\title{sm2vec.tools}
\usage{
sm2vec.tools(m, diag = FALSE)
}
\arguments{
\item{m}{symmetric matrix}

\item{diag}{include the main diagonal}
}
\value{
vector
}
\description{
Symm matrix to a vector
}
\examples{

}
\references{
\url{https://doi.org/10.2202/1544-6115.1175}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
