% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2smTools.R
\docType{package}
\name{vec2sm.tools}
\alias{vec2sm.tools}
\title{vec2sm.tools}
\usage{
vec2sm.tools(vec)
}
\arguments{
\item{vec}{vector}
}
\value{
symmetric matrix
}
\description{
Symm matrix to a vector
}
\examples{

}
\references{
\url{https://doi.org/10.2202/1544-6115.1175}
}
\author{
Victor Bernal \email{v.a.bernal.arzola@rug.nl; victor.arturo.bernal@gmail.com}
}
